Refactor GMX testing into package and add GmxLib validation docs
- Move eth_defi/gmx/testing.py into eth_defi/gmx/testing/ package with
  constants, fork_provider, keeper, and oracle submodules
- Add GmxLib validation details section to README-GMX-Lagoon.md
- Add irregular_reporting to BAD_FLAGS in vault/flag.py
- Update tests/gmx/fork_helpers.py to re-export from new package
